+++ server/modules/defaults/src/lib.rs
@@ -49,7 +49,8 @@ use {
     reovim_module_explorer as explorer, reovim_module_keymap as keymap, reovim_module_lsp as lsp,
     reovim_module_lsp_navigation as lsp_navigation, reovim_module_microscope as microscope,
     reovim_module_motions as motions, reovim_module_notification as notification,
-    reovim_module_range_finder as range_finder, reovim_module_scratch_buffer as scratch_buffer,
+    reovim_module_profiles as profiles, reovim_module_range_finder as range_finder,
+    reovim_module_scratch_buffer as scratch_buffer,
     reovim_module_search as search, reovim_module_snippet as snippet,
     reovim_module_tetromino as tetromino, reovim_module_treesitter_markdown as treesitter_markdown,
     reovim_module_treesitter_rust as treesitter_rust, reovim_module_undo as undo,
@@ -150,6 +151,8 @@ impl DefaultsModule {
             Box::new(vim_explorer::VimExplorerModule::new()),
             // Tetromino game (#537)
             Box::new(tetromino::TetrominoModule::new()),
+            // Configuration profiles (#593)
+            Box::new(profiles::ProfilesModule::new()),
         ]
     }
 }
@@ -227,6 +230,8 @@ impl Module for DefaultsModule {
             ModuleId::new("vim-microscope"),
             // Tetromino game (#537)
             ModuleId::new("tetromino"),
+            // Configuration profiles (#593)
+            ModuleId::new("profiles"),
         ]
     }